Fix settings with complicated values containing =
The older code was splitting on the = and using the last setion of the
datagram as the setting value, which means any value containing an
equals sign was losing everything but the first part of the setting.

When those settings were then saved again, they were written to the
settings file on the ground station with just the last part of the
setting value, which broke the config file.