Allow widgets to declare their own Popups
These can be completely controlled by the widget itself, rather than
positioned automatically near the widget and given the translucent
overlay look.

To use one, add hasWidgetPopup: true and widgetPopup: Popup {} in the
widget itself to override the properties in BaseWidet. Tapping the
widget will then open the popup.

It is not possible to do both a full popup and the detail panel at the
moment so the detail panel will be ignored if one is defined in the
widget.