Add network link between app on Android<->Ground
Currently has support for syncing widget movements from Android to
the ground side, for users who have a ground station with a screen
but no touchscreen support to move widgets directly. This makes it
unnecessary for them to edit files manually just to change the widget
locations.

The link will also be used for things like sending cached map tiles
to the ground station, which may not have internet access to fetch
the tiles directly.

Currently the link operates one-way from Android->ground but only
because of the hardcoded connect address, once mDNS is added it will
be a 2-way link where both sides are just peers.