Switch Windows builds to 32-bit for now
Long term there should be a separate build script that takes the
output of 32-bit and 64-bit builds and packages them together. NSIS
supports this and would make it unnecessary for uses to choose or care
whether they need 32-bit or 64-bit builds.

The 64-bit builds will likely help when software h264 decoding is in
use, otherwise it may not make much difference.