Change hardware decoder switch to "force software"
The UI can't claim to enable hardware decoding with a switch, that
would be misleading.

The pipeline construction and the different decoder elements for every
hardware decoder on every platform are too complex to handle without
using decodebin3 to construct a working pipeline automatically and
hope that it finds a usable hardware decoder.

However, the UI can promise to force software decoding, because we can
easily construct a software-only pipeline.