Add warning watermark on main video
This is to warn the user that flying with the app is not recommended
yet due to video framerate and jitter bugs that need to be worked out.

This warning only appears when the app is handling the main video, for
example on Android, but not on the ground station because hello_video
is being used at the moment.