Fix drag-and-drop on ground station
This is a quick fix that allows drag-and-drop to begin working again
on the ground station touchscreen without breaking remote
drag-and-drop.

This part of the code sends drag messages over the network, something
that is only intended to work from android->ground, not from
ground->anything.

The current issue is that the drag messages are being sent by
the ground station to itself, which prevents locally dragging from
working properly.

Long term it would be wise to fix the drag-and-drop system so that it
is 2-way, which would necessarily handle issues like this anyway.