Don't save settings unless their value has changed
This replaces the previous simplistic code that just set a modified
property on the setting in the model and used that to detect which
settings had actually changed.

The problem is that for various reasons that flag gets set even when
the setting value isn't actually different from the original value,
presumably because the onValueChanged callback in QML gets triggered
for other reasons.

Instead we store the original setting value and check it during a save
which should prevent any setting from being saved unless it actually
changed.

Fixes #55