Add telemetry charts in status panel
There is no defined time scale for the X axis, but for the moment
that is OK as the charts are primarily intended to show a recent
snapshot rather than a specific time period. We can add full time
support later on.

These may have an effect on rendering which would require disabling
them when they aren't visible.