Move chart to separate element w/loader to disable
The charts are not exactly stable at the moment, the plots disappear
when the settings panel is closed and reopened and the opengl render
option for the plots may be causing strange rendering issues. They
will be disabled on stable releases for now, might add an option
to enable them in settings area.