Temporarily disable horizon widget size control
The detail panel is automatically shown on the corner of the screen
closest to the widget, but resizing the widget can suddenly cause that
calculation to change while the panel is still open, making the panel
jump across the screen.

It will likely need to be in the drag-and-drop panel.