Add video decoder and rendering layer for MMAL/Pi
This uses the MMAL API to decode incoming h264 NAL units, then sends
them to a standard QML VideoOutput rendering system based on
QVideoFrame.

At the moment there is an extra memcpy() step, for PiP it won't be
an issue but for 1080p high framerate it might. There is a way to
feed the decoded buffers directly into a QAbstractVideoBuffer, either
with raw YUV data or a GL texture, code to handle that is being
tested separately.