Add missing privacy string required on iOS
QtMultimedia links against the code that can control the camera, so
the privacy string is required by apple.

Ordinarily if we were using the camera, this string would be shown to
users when the app asks for permission to use it, but we don't so they
will never see it.