Add support for using a handle to unlock dragging
This allows widgets to declare that they want a small hand icon to
stay in the corner of the widget and be used to initiate drag-and-drop
instead of the entire widget, since some widgets need to handle touch
events themselves.

This is only used by the map right now and may not be useful anywhere
else.