Merge the large and small maps, move map settings
This changes the way maps work so that there is just one map now,
and the map can be made full screen and collapsed again using the
button in the corner.

Settings are moved from the detail panel to the slider in the map
itself when it is expanded fullscreen. We can partially revert that
and put the settings both places if we just put another small button
in the corner of the map to trigger the detail panel again.

This also enables the drag handle in the lower right corner of the map
so that touches can be handled by the map itself without breaking
drag-and-drop.

The circle map type is currently disabled until we can figure out a
way to allow that shape without giving up the corners and edges, which
are one of the few ways we can add interactions for things like
drag-and-drop, resizing, etc, while allowing the map itself to remain
interactive even when collapsed.