Use fewer but larger MMAL input buffers
30x 1080p input buffer frames is a lot, the decoder complains about
it sometimes.

However we don't really want the input side to be using a specific
size chosen by the decoder during configuration, we want to ensure
that the buffer sizes are always large enough for even 1080p video
in order to avoid sudden invalid memory access when new NAL units
come in during reconfiguration (there is currently no locking when
that happens).