Convert the settings Popup to a Rectangle
When running at 1920x1080 on a Pi 3B, the Popup for some reason causes
Qt to allocate GL textures larger than the GPU can handle, which then
causes GL calls to fail, leading to rendering bugs.

It may be that Popup is not intended to be used fullscreen like that.