Automatically trigger ground IP event when on Pi
Ordinarily, when running on Android or other platforms, as soon as
the device joins the hotspot QOpenHD would begin receiving packets
from the settings system on the ground station, triggering the events
that tell the rest of the app what the ground station IP is so they
can begin doing things like connect to the microservice router.

When running on the ground station itself, that doesn't occur until
something sends a message to the settings system. So we trigger that
event manually, only once at launch, in order to make everything work
the same as it would on other platforms.