Add StatusView for air/ground logs
Used to display logs of Open.HD setup and configuration all the way
back to early boot. Messages are stored by the status microservice,
then once QOpenHD is running it can request all of them from both
the air and ground side.

Messages should be sorted by timestamp, this is not currently done.