Add mavlink sysid setting
Uses different default sysid for different platforms as a simple way
to ensure that someone using a ground station and also android won't
have to immediately change the sysid on one of them. People using 2x
of the same platform will have to change one of them, but we can move
to randomizing them if needed.