Make the bitrate widget larger, occupy whole column
This allows the bitrate widget to have detail text under it without
overlapping the rest of the OSD. Previously it would have conflicted
with the location of the downlink RSSI widget (which also has detail
text that can be enabled under it).

We may have to make these resizable as people will have preferences
for one or the other style.