Make the downlink RSSI widget larger, whole column
This allows the downlink RSSI widget to have detail text under it
without overlapping the rest of the OSD. Previously it would have
conflicted with the location of the bitrate widget (which also has
detail text that can be enabled under it).

We may have to make these resizable as people will have preferences
for one or the other style.