Add native video decode/render on Android
Needs a few fixes before it can be enabled by default, the 3rd party
h264 parser is not very well written and should be replaced entirely,
as it has a lot of memory safety bugs that can lead to a crash.

The renderer also needs a few tweaks for performance.