Handle RTP FU_A packet loss by discarding the rest
When an FU_A NAL is arriving, we should see the header, one or more
regular pieces, and then an end. If we never see the end but the
flag is still set, we know that some of the pieces did not arrive.

There is likely a better way to handle this.