Allow stereo to support different modes
Mode 0 is disabled
Mode 1 is Aspect-fit

Some others will likely be a square in the center of the 2 viewing
areas to keep widgets away from the outer edges, and then probably
a mode that does stretch fit but does not attempt to maintain the
aspect ratio, so that users can position and resize things however
they want. We will have to find a good way to store widgets positioned
in different modes in the near future.