Fix clickthrough from settings panel to OSD widgets
When the settings panel is up, for some reason Qt is allowing clicks
to go through to the OSD widgets even though the entire screen is
covered and the settings panel is not transparent.

Adding a fake mouse area and preventing events from propagating back
to the layers below it prevents those clicks from happening.