Add comment with approximate cpu usage of widgets
This is not exact, as it's difficult to isolate them from each other,
but it should be fairly close.

Testing was done by enabling only the declarative widgets, then
enabling the others one at a time to measure CPU usage, both overall
and in the render and main threads.

In some cases, higher cpu use alone doesn't tell the whole story,
because using a lot of blocking calls in the main thread can cause
more problems than simply using a lot of CPU time rendering
or uploading GPU textures.

Canvas is not the only issue here, but it does seem to hit fairly
hard. In some of these, animations use a surprising amount of CPU
time, and in some of them Shape or translate might as well.

The ones using a lot of CPU time will be ported to OpenGL which should
drop them down to almost nothing.