Allow multiple sysid/compid mavlink messages
This is a crude hack but fine for now, it allows us to filter the
incoming mavlink messages according to the target device we're
listening for.

It's not the right way to handle the problem but doesn't cause any
problems for now, will be replaced once we're sending other mavlink
messages.