Fix RC RSSI showing last value when messages stop
When the RC radio is turned off or for some other reason the telemetry
messages stop, the RSSI widget should not continue showing the same
RSSI value, it should indicate that there is no longer a signal.

A timer keeps track of the last time the message arrived and resets
the value to 0 if it's been more than 3 seconds.