Force save the widget centering settings on load
This ensures that they're actually present in the settings system
when something tries to load all of the key/value pairs, without this
any setting that hasn't been explicitly saved seems to be absent.

And that causes serious problems if the goal is to retrieve all of the
settings and save them to a file.