Add settings file template
This must be kept updated with any settings that are added in the app,
if not they will simply never end up in the saved file and people
will not be able to change them.

If any settings are *removed* from the app but left in the template,
the template engine will fail to render the file, because it expects
to find all of the template values and currently we are bulk loading
the entire key/value pair list from Settings to avoid manually loading
them all one by one, by name.

This is fragile, but it's the best way to make a settings file that
can be auto-updated with new comments/instructions and new settings,
it ensures that users don't have to manually update the file by
searching through 600+ lines to find out what changed.