Retry saving settings file at boot on pi
There is a race condition in the Qt settings system, when it loads
the Settings element with all the defaults it will merge the defaults
with the changed settings already saved on disk, and then save all
of them to the internal settings file in the user folder.

However there's a delay there, Qt will group changed settings together
to avoid re-saving the file hundreds of times for no reason. The
delay is 500ms, and this is just enough to cause a problem at boot,
because we assume that having the Settings element signal it has
loaded means the settings are now all ready, but that's not the case,
it can be several seconds later. If we try just once, our custom
settings file system will simply fail, and until now it was also
throwing an exception which was not being caught.

We now retry for 10 seconds, which should be more than enough and will
deal with the race condition in Qt.