Fix language change causing settings row glitches
The individual panels of each settings tab were appearing on top of
each other when the language was changed. Still don't know why it
happens but this fixes the issue, despite being redundant since this
is a StackLayout that should be hiding the panels when they aren't
supposed to be visible anyway.