Show OSDCustomizer when tap-and-hold screen
This reuses the stereo mode exit touch area for showing the osd
customizer, you tap and hold any area between widgets and it will
show up. Then you can tap outside it to dismiss it again.

This reuses the font selection element, but the code for the color
selection is slightly different from the app settings panel. We may
still be able to make them a reusable element but this is OK for now.