Replace ColorDialog with the ColorPicker element
This commit only changes the OSD customizer, the settings area
still needs to be switched over to use this element instead.

The ColorDialog clearly doesn't work on the Pi at the moment, and
that might be our fault (none of the Dialog types work right), but
we don't have time to mess with it and this new type works better
and consistently on every platform anyway.