Allow triple-tap on video area to enable stereo
This makes turning it on and off easier than going into the settings
area, and should not conflict with anything else. No other UI
interaction requires 3 taps, and when stereo mode is off the rest
of the UI should be in "front" of this touch area. When stereo mode is
on, the touch area that turns it back off again is in front of
everything else.