Dynamically set ADSB ground address
This is a callback that comes from the settings class, which knows
where the ground station is because the ground station sends GroundIP
messages to every hotspot device periodically.

We're moving away from that settings system and toward mavlink params
but this same idea will still work, it'll just be a standard heartbeat
instead of a custom packet.